Update edmonton_weather_raw_data.csv
I have now proceeded with the pre-processing of this data. I have removed attributes that I have no interest in investigating for the purposes of this project.

I am also only interested in the dates of March 11th 2018 to March 10th 2020 as this was the time period I was staying in Edmonton.

I also made sure that any value that was empty was instead replaced with the value "0".